rename fhir utils to mapping constants
